global-server/pyproject.toml
[tool.poetry]
name = "restaraunt-petproject/global-server"
version = "0.1.0"
description = ""
authors = ["Vchslv-Zkhrv <vchslv.zkhrv@gmail.com>"]
packages = [{include = "restaraunt_petproject"}]

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.1.1"
fastapi = "0.101.0"
loguru = "^0.7.2"
passlib = "^1.7.4"
colorama = "^0.4.6"
python-dotenv = "^1.0.0"
asyncpg = "^0.29.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
typeguard = "^4.1.5"
email-validator = "^2.1.0.post1"
python-ulid = "^2.2.0"
pydantic-settings = "^2.2.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
flake8 = "^6.1.0"

[tool.poetry.group.test.dependencies]
pytest-dotenv = "^0.5.2"
pytest-asyncio = "^0.23.5.post1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
